feat: OeX_ES-12: updates versions of the ES and the ES DSL libs
changes ES test environment from V1.5.2 to V7.8.0

rewrites index documents for compatibility with ES7

updates mappings creation of ES documents

updates tests and document data serializer

edit documents mappping and query methods

update travs-ci file and make commands for tests running

update requirements

modify cohort field type in the learner serializer

removes pip-tools library version constraint, updates requirements

edits README description

updates README description

updates travis.ci

rebase with master
