+++ analytics_data_api/v0/tests/views/test_learners.py
@@ -94,9 +94,8 @@ class LearnerAPITestMixin(CsvViewMixin):
             if optional_field[1]:
                 body[optional_field[0]] = optional_field[1]
 
-        self._es.create(
+        self._es.index(
             index=settings.ELASTICSEARCH_LEARNERS_INDEX,
-            doc_type='roster_entry',
             body=body
         )
 
@@ -116,9 +115,8 @@ class LearnerAPITestMixin(CsvViewMixin):
         """
         Created an index with the date of when the learner index was updated.
         """
-        self._es.create(
+        self._es.index(
             index=settings.ELASTICSEARCH_LEARNERS_UPDATE_INDEX,
-            doc_type='marker',
             body={
                 'date': date,
                 'target_index': settings.ELASTICSEARCH_LEARNERS_INDEX,
@@ -224,7 +222,7 @@ class LearnerTests(VerifyCourseIdMixin, LearnerAPITestMixin, TestCaseWithAuthent
         }
         self.assertDictEqual(expected, response.data)
 
-    @patch('analytics_data_api.v0.models.RosterEntry.get_course_user', Mock(return_value=[]))
+    @patch('analytics_data_api.v0.documents.RosterEntry.get_course_user', Mock(return_value=[]))
     def test_not_found(self):
         user_name = 'a_user'
         course_id = 'edX/DemoX/Demo_Course'
