+++ analyticsdataserver/settings/base.py
@@ -95,6 +95,12 @@ ELASTICSEARCH_AWS_SECRET_ACCESS_KEY = None
 ELASTICSEARCH_CONNECTION_CLASS = None
 # only needed with BotoHttpConnection, e.g. 'us-east-1'
 ELASTICSEARCH_CONNECTION_DEFAULT_REGION = None
+
+DEFAULT_ELASTICSEARCH_INDEX_SETTINGS = {
+    'number_of_shards': 1,
+    'number_of_replicas': 0
+}
+ELASTICSEARCH_INDEX_SETTINGS = environ.get('ELASTICSEARCH_INDEX_SETTINGS', DEFAULT_ELASTICSEARCH_INDEX_SETTINGS)
 ########## END ELASTICSEARCH CONFIGURATION
 
 ########## GENERAL CONFIGURATION
