+++ requirements/constraints.txt
@@ -16,8 +16,8 @@
 # django-storages version 1.9 drops support for boto storage backend.
 django-storages<1.9                 # BSD
 
-# elasticsearch-dsl depends on elasticsearch >2.0.0,<3.0.0
-elasticsearch-dsl<2.0.0
+# elasticsearch-dsl depends on elasticsearch >=7.8.0,<8.0.0
+elasticsearch-dsl>=7.2.1,<8.0.0
 
 # Python3.5 compatibility is broken until this
 # gets merged https://github.com/LuminosoInsight/ordered-set/pull/68
