+++ requirements/doc.txt
@@ -34,8 +34,8 @@ edx-enterprise-data==2.2.8  # via -r requirements/base.in
 edx-opaque-keys==2.1.1    # via -r requirements/base.in, edx-ccx-keys, edx-drf-extensions, edx-enterprise-data
 edx-rbac==1.3.4           # via edx-enterprise-data
 edx-rest-api-client==5.3.0  # via -r requirements/base.in, edx-enterprise-data
-elasticsearch-dsl==0.0.11  # via -c requirements/constraints.txt, -r requirements/base.in
-elasticsearch==1.9.0      # via elasticsearch-dsl
+elasticsearch-dsl==7.2.1  # via -c requirements/constraints.txt, -r requirements/base.in
+elasticsearch==7.8.0     # via elasticsearch-dsl
 future==0.18.2            # via pyjwkest
 idna==2.10                # via requests
 itypes==1.2.0             # via coreapi
