+++ requirements/pip_tools.in
@@ -1,3 +1,3 @@
 # Dependencies to run compile tools
-pip-tools>5.0.0                           # Contains pip-compile, used to generate pip requirements files
+pip-tools                          # Contains pip-compile, used to generate pip requirements files
 six
